Move join_paths from core.util to core.path
Move its tests, too.
