+++ pyanaconda/modules/common/structures/payload.py
@@ -21,7 +21,7 @@
 from dasbus.structure import DBusData
 from dasbus.typing import *  # pylint: disable=wildcard-import
 
-from pyanaconda.core.util import join_paths
+from pyanaconda.core.path import join_paths
 from pyanaconda.core.constants import URL_TYPE_BASEURL, DNF_DEFAULT_REPO_COST
 
 __all__ = ["RepoConfigurationData", "SSLConfigurationData"]
