+++ pyanaconda/modules/network/installation.py
@@ -18,8 +18,7 @@
 import os
 import shutil
 
-from pyanaconda.core import util
-from pyanaconda.core.path import make_directories
+from pyanaconda.core.path import make_directories, join_paths
 from pyanaconda.modules.common.errors.installation import NetworkInstallationError
 from pyanaconda.modules.common.task import Task
 from pyanaconda.anaconda_loggers import get_module_logger
@@ -184,14 +183,14 @@ Name={}
         """
 
         if ifname_option_values:
-            target_system_dir = util.join_paths(root, self.SYSTEMD_NETWORK_CONFIG_DIR)
+            target_system_dir = join_paths(root, self.SYSTEMD_NETWORK_CONFIG_DIR)
             make_directories(target_system_dir)
 
         for ifname_value in ifname_option_values:
             iface, mac = ifname_value.split(":", 1)
             content = self.INTERFACE_RENAME_FILE_CONTENT_TEMPLATE.format(mac, iface)
             config_file = self.INTERFACE_RENAME_FILE_TEMPLATE.format(iface)
-            config_file_path = util.join_paths(self.SYSTEMD_NETWORK_CONFIG_DIR, config_file)
+            config_file_path = join_paths(self.SYSTEMD_NETWORK_CONFIG_DIR, config_file)
             _write_config_file(
                 root,
                 config_file_path,
