+++ pyanaconda/modules/subscription/installation.py
@@ -23,7 +23,7 @@ from dasbus.typing import get_variant, Str
 
 from pyanaconda.core import util
 from pyanaconda.core.constants import RHSM_SYSPURPOSE_FILE_PATH
-from pyanaconda.core.path import make_directories
+from pyanaconda.core.path import make_directories, join_paths
 from pyanaconda.core.subscription import check_system_purpose_set
 
 from pyanaconda.modules.common.task import Task
@@ -68,7 +68,7 @@ class ConnectToInsightsTask(Task):
                       "Insights requested but target system is not subscribed, skipping")
             return
 
-        insights_path = util.join_paths(self._sysroot, self.INSIGHTS_TOOL_PATH)
+        insights_path = join_paths(self._sysroot, self.INSIGHTS_TOOL_PATH)
         # check the insights client utility is available
         if not os.path.isfile(insights_path):
             raise InsightsClientMissingError(
@@ -190,7 +190,7 @@ class TransferSubscriptionTokensTask(Task):
     def _transfer_file(self, target_path, target_name):
         """Transfer a file with nice logs and raise an exception if it does not exist."""
         log.debug("subscription: transferring %s", target_name)
-        target_repo_file_path = util.join_paths(self._sysroot, target_path)
+        target_repo_file_path = join_paths(self._sysroot, target_path)
         if not self._copy_file(target_path, target_repo_file_path):
             msg = "{} ({}) is missing".format(target_name, self.RHSM_REPO_FILE_PATH)
             raise SubscriptionTokenTransferError(msg)
